+++ src/utils/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_DURATION = 2
+export const DEFAULT_START = 0