+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export { easings, getEasing, defaultEasing } from './easing'
 export { getToLocaleStringParamsSupport } from './toLocaleStringSupport'
+export { DEFAULT_DURATION, DEFAULT_START } from './constants'